Refactor GetUntiScreen component: Simplify flow by removing EML upload and ZK proof generation steps, update UI for email sent confirmation
